refactor: clean up openai proxy router code
