+++ api/core/domain/tenant_data.py
@@ -32,6 +32,16 @@ class PublicOrganizationData(BaseModel):
     def is_anonymous(self) -> bool:
         return not self.org_id and not self.owner_id
 
+    @property
+    def base_app_url(self) -> str:
+        return f"{WORKFLOWAI_APP_URL}/{self.slug}"
+
+    def app_schema_url(self, agent_id: str, schema_id: int):
+        return f"{self.base_app_url}/agents/{agent_id}/{schema_id}"
+
+    def app_deployments_url(self, agent_id: str, schema_id: int):
+        return f"{self.app_schema_url(agent_id, schema_id)}/deployments"
+
 
 class TenantData(PublicOrganizationData):
     anonymous_user_id: str | None = Field(default=None)
@@ -76,13 +86,3 @@ class TenantData(PublicOrganizationData):
             return False
 
         return True
-
-    @property
-    def base_app_url(self) -> str:
-        return f"{WORKFLOWAI_APP_URL}/{self.slug}"
-
-    def app_schema_url(self, agent_id: str, schema_id: int):
-        return f"{self.base_app_url}/agents/{agent_id}/{schema_id}"
-
-    def app_deployments_url(self, agent_id: str, schema_id: int):
-        return f"{self.app_schema_url(agent_id, schema_id)}/deployments"
