+++ api/core/storage/__init__.py
@@ -6,12 +6,21 @@ from core.domain.errors import InternalError
 
 # TODO: duplicate of ObjectNotFoundError, we should use only one of them
 class ObjectNotFoundException(InternalError):
-    code: ErrorCode = "object_not_found"
+    default_code: ErrorCode = "object_not_found"
 
-    def __init__(self, msg: str | None = None, code: ErrorCode | None = None, **extras: Any):
-        super().__init__(msg, **extras)
-        self.code = code or self.code
-        self.extras = extras or {}
+    def __init__(
+        self,
+        msg: str | None = None,
+        code: ErrorCode | None = None,
+        object_type: str | None = None,
+        **extras: Any,
+    ):
+        super().__init__(msg, object_type=object_type, **extras)
+        self.code: ErrorCode = code or self.default_code
+
+    @property
+    def object_type(self) -> str | None:
+        return self.extras.get("object_type")
 
 
 # TODO[ids]: passing as a tuple for now to reduce the amount of changes needed
